server/pyproject.toml
[project]
name = "memlog"
version = "0.0.1"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.118",
    "whoosh>=2.7",
    "python-jose[cryptography]>=3.3",
    "pyotp>=2.9",
    "qrcode>=8.0",
    "aiofiles>=24.1",
    "python-multipart>=0.0.20",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.8",
    "mypy>=1.11",
    "types-aiofiles>=24.1",
    "types-python-jose>=3.3",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
